fix interaction between Task & Task2; add python.Placeholder
Any OnTask plugin that changes Task2.request has no effect because that
field is not taken into account in Task2.statements.
Fixing this by introducing python.Placeholder, a syntax node
representing a non-syntax node (like a Request instance, in our case).
Special care has to be taken when the Task being converted into a Task2
already has a Task.locust_request, which is kind of a copy of
Task.request with a different interface.

Fix #31.

Signed-off-by: Thibaut Le Page <thibaut.le.page@zalando.de>